% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocode-package.R
\docType{package}
\name{arcgisgeocode-package}
\alias{arcgisgeocode}
\alias{arcgisgeocode-package}
\title{arcgisgeocode: A Robust Interface to ArcGIS 'Geocoding Services'}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A very fast and robust interface to ArcGIS 'Geocoding Services'. Provides capabilities for reverse geocoding, finding address candidates, character-by-character search autosuggestion, and batch geocoding. The public 'ArcGIS World Geocoder' is accessible for free use via 'arcgisgeocode' for all services except batch geocoding. 'arcgisgeocode' also integrates with 'arcgisutils' to provide access to custom locators or private 'ArcGIS World Geocoder' hosted on 'ArcGIS Enterprise'. Learn more in the 'Geocode service' API reference \url{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-arcgis/arcgisgeocode}
  \item \url{https://developers.arcgis.com/r-bridge/api-reference/arcgisgeocode}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})
}

}
\keyword{internal}
