% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-geocode-job.R
\name{BatchGeocodeJob}
\alias{BatchGeocodeJob}
\title{Batch Geocode Job}
\description{
A \code{\link[arcgisutils:arc_gp_job]{arcgisutils::arc_gp_job}} for the \code{BatchGeocode} endpoint, which takes a
JSON body and reports status over \code{POST}.
}
\section{Super class}{
\code{\link[arcgisutils:arc_gp_job]{arcgisutils::arc_gp_job}} -> \code{BatchGeocodeJob}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{item_id}}{the portal item holding the uploaded addresses.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{status}}{the job status as an \code{\link[arcgisutils:arc_job_status]{arcgisutils::arc_job_status()}}.}

    \item{\code{results}}{the URL of the geocoded file.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BatchGeocodeJob-initialize}{\code{BatchGeocodeJob$new()}}
    \item \href{#method-BatchGeocodeJob-start}{\code{BatchGeocodeJob$start()}}
    \item \href{#method-BatchGeocodeJob-write}{\code{BatchGeocodeJob$write()}}
    \item \href{#method-BatchGeocodeJob-clone}{\code{BatchGeocodeJob$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="arcgisutils" data-topic="arc_gp_job" data-id="await"><a href='../../arcgisutils/html/arc_gp_job.html#method-arc_gp_job-await'><code>arcgisutils::arc_gp_job$await()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="arcgisutils" data-topic="arc_gp_job" data-id="cancel"><a href='../../arcgisutils/html/arc_gp_job.html#method-arc_gp_job-cancel'><code>arcgisutils::arc_gp_job$cancel()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="arcgisutils" data-topic="arc_gp_job" data-id="messages"><a href='../../arcgisutils/html/arc_gp_job.html#method-arc_gp_job-messages'><code>arcgisutils::arc_gp_job$messages()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchGeocodeJob-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BatchGeocodeJob-initialize}{}}}
\subsection{\code{BatchGeocodeJob$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BatchGeocodeJob$new(base_url, item_id, field_mapping, token = arc_token())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{base_url}}{the URL of the job service, without \verb{/submitJob}.}
      \item{\code{item_id}}{the id of the portal item holding the addresses.}
      \item{\code{field_mapping}}{a named character vector of geocoder field to column.}
      \item{\code{token}}{an \code{\link[arcgisutils:arc_token]{arcgisutils::arc_token()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchGeocodeJob-start"></a>}}
\if{latex}{\out{\hypertarget{method-BatchGeocodeJob-start}{}}}
\subsection{\code{BatchGeocodeJob$start()}}{
  Submits the job. The endpoint requires a JSON body.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BatchGeocodeJob$start()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchGeocodeJob-write"></a>}}
\if{latex}{\out{\hypertarget{method-BatchGeocodeJob-write}{}}}
\subsection{\code{BatchGeocodeJob$write()}}{
  Downloads the geocoded file.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BatchGeocodeJob$write(path)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{where to write the result CSV.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchGeocodeJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchGeocodeJob-clone}{}}}
\subsection{\code{BatchGeocodeJob$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BatchGeocodeJob$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
