% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{esri_wkids}
\alias{esri_wkids}
\title{Esri well-known IDs}
\format{
An object of class \code{integer} of length 2886.
}
\usage{
esri_wkids
}
\value{
a numeric vector of well-known IDs
}
\description{
An integer vector containing the WKIDs of Esri authority
spatial references.
Esri WKIDs were identified from the \href{https://cran.r-project.org/package=arcgeocoder}{\code{{arcgeocoder}}} package from
\href{https://github.com/dieghernan}{@dieghernan}.
}
\keyword{datasets}
