% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder-obj.R
\name{geocode_server}
\alias{geocode_server}
\title{Create a GeocodeServer}
\usage{
geocode_server(url = world_geocoder_url)
}
\arguments{
\item{url}{the url}
}
\description{
Create an object of class \code{GeocodeServer} from a URL. This object
stores the service definition of the geocoding service as a list object.
}
