% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-iso-3166.R
\name{iso_3166_codes}
\alias{iso_3166_codes}
\title{ISO 3166 Country Codes}
\usage{
iso_3166_codes()
}
\value{
a \code{data.frame} with columns \code{country}, \code{code_2}, \code{code_3}.
}
\description{
Create a data.frame of ISO 3166 2 and 3 digit Country codes.
}
\details{
Country codes provided by \href{https://docs.rs/rust_iso3166/latest/rust_iso3166/index.html}{\code{rust_iso3166}}.
}
\examples{
head(iso_3166_codes())
}
