% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocode-package.R
\name{storage}
\alias{storage}
\title{Storing Geocoding Results}
\description{
The results of geocoding operations cannot be stored or
persisted unless the \code{for_storage} argument is set to
\code{TRUE}. The default argument value is \code{for_storage = FALSE}, which indicates the results of the operation can't be stored, but they can be temporarily displayed on a map, for instance. If you store the results, in a database, for example, you need to set this parameter to true.
}
\details{
See \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-find-address-candidates.htm#ESRI_SECTION3_BBCB5704B46B4CDF8377749B873B1A7F}{the official documentation} for more context.
}
