% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest.R
\name{suggest_places}
\alias{suggest_places}
\title{Search Suggestion}
\usage{
suggest_places(
  text,
  location = NULL,
  category = NULL,
  search_extent = NULL,
  max_suggestions = NULL,
  country_code = NULL,
  preferred_label_values = NULL,
  geocoder = default_geocoder(),
  token = arc_token()
)
}
\arguments{
\item{preferred_label_values}{default NULL. Must be one of \code{"postalCity"}
or \code{"localCity"}.}

\item{geocoder}{the url to the geocoder service.
Default provided by \code{default_geocoder()}.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[arcgisutils:auth_code]{auth_code()}}
or related function}
}
\value{
A \code{data.frame} with 3 columns: \code{text}, \code{magic_key}, and \code{is_collection}.
}
\description{
Search Suggestion
}
\details{
Unlike the other functions in this package, \code{suggest_places()} is not
vectorized as it is intended to provide search suggestions for individual
queries such as those made in a search bar.
}
\examples{
# identify a search point
location <- sf::st_sfc(sf::st_point(c(-84.34, 33.74)), crs = 4326)

# create a search extent from it
search_extent <- sf::st_bbox(sf::st_buffer(location, 10))

# find suggestions from it
suggestions <- suggest_places(
  "bellwood",
  location,
  search_extent = search_extent
)

# get address candidate information
# using the text and the magic key
find_address_candidates(
  suggestions$text,
  magic_key = suggestions$magic_key
)
}
