% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{world_geocoder}
\alias{world_geocoder}
\title{ArcGIS World Geocoder}
\format{
An object of class \code{GeocodeServer} (inherits from \code{list}) of length 12.
}
\usage{
world_geocoder
}
\value{
an object of class \code{GeocodeServer}
}
\description{
The \href{https://www.esri.com/en-us/arcgis/products/arcgis-world-geocoder}{ArcGIS World Geocoder}
is made publicly available for some uses. The \code{world_geocoder} object is used
as the default \code{GeocodeServer} object in \code{\link[=default_geocoder]{default_geocoder()}} when no
authorization token is found. The \code{\link[=find_address_candidates]{find_address_candidates()}},
\code{\link[=reverse_geocode]{reverse_geocode()}}, and \code{\link[=suggest_places]{suggest_places()}} can be used without an
authorization token. The \code{\link[=geocode_addresses]{geocode_addresses()}} funciton requires an
authorization token to be used for batch geocoding.
}
\keyword{datasets}
