% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-geocode-job.R
\name{write_geocode_result}
\alias{write_geocode_result}
\title{Write Batch Geocoding Results}
\usage{
write_geocode_result(job, path)
}
\arguments{
\item{job}{A \code{BatchGeocodeJob}.}

\item{path}{String. Where to write the result CSV.}
}
\value{
\code{path}, invisibly.
}
\description{
Downloads the result of a completed \code{\link[=geocode_file_job]{geocode_file_job()}}.
}
\examples{
\dontrun{
write_geocode_result(job, "geocoded.csv")
}
}
