% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocode-package.R, R/extendr-wrappers.R
\docType{package}
\name{arcgisgeocode-package}
\alias{arcgisgeocode-package}
\title{arcgisgeocode: A Robust Interface to ArcGIS Geocode Services}
\description{
A very fast and robust interface to 'ArcGIS Geocode Services.' Provides capabilities to reverse geocode, find address candidates, search autosuggestion, and batch geocoding. The public 'ArcGIS World Geocoder' is accessible to the public via 'arcgisgeocode' for all services except batch geocoding. 'arcgisgeocode' also integrates with 'arcgisutils' to provide access to private custom locators or 'ArcGIS World Geocoder's.

A very fast and robust interface to 'ArcGIS Geocode Services.' Provides capabilities to reverse geocode, find address candidates, search autosuggestion, and batch geocoding. The public 'ArcGIS World Geocoder' is accessible to the public via 'arcgisgeocode' for all services except batch geocoding. 'arcgisgeocode' also integrates with 'arcgisutils' to provide access to private custom locators or 'ArcGIS World Geocoder's.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-arcgis/arcgisgeocode}
  \item \url{https://r.esri.com/arcgisgeocode}
}


Useful links:
\itemize{
  \item \url{https://github.com/r-arcgis/arcgisgeocode}
  \item \url{https://r.esri.com/arcgisgeocode}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
