% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-candidates.R
\name{find_address_candidates}
\alias{find_address_candidates}
\title{Find Address Candidates}
\usage{
find_address_candidates(
  single_line = NULL,
  address = NULL,
  address2 = NULL,
  address3 = NULL,
  neighborhood = NULL,
  city = NULL,
  subregion = NULL,
  region = NULL,
  postal = NULL,
  postal_ext = NULL,
  country_code = NULL,
  search_extent = NULL,
  location = NULL,
  category = NULL,
  crs = NULL,
  max_locations = NULL,
  for_storage = FALSE,
  match_out_of_range = NULL,
  location_type = NULL,
  lang_code = NULL,
  source_country = NULL,
  preferred_label_values = NULL,
  magic_key = NULL,
  geocoder = default_geocoder(),
  token = arc_token(),
  .progress = TRUE
)
}
\arguments{
\item{crs}{the CRS of the returned geometries. Passed to \code{sf::st_crs()}.}

\item{for_storage}{default \code{FALSE}. Whether or not the results will be saved
for long term storage. See Details.}

\item{location_type}{default \code{"rooftop"}. Must be one of \code{"rooftop"} or \code{"street"}.}

\item{preferred_label_values}{default NULL. Must be one of \code{"postalCity"}
or \code{"localCity"}.}

\item{geocoder}{the url to the geocoder service.
Default provided by \code{default_geocoder()}.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[arcgisutils:auth_code]{auth_code()}}
or related function}

\item{.progress}{default \code{TRUE}. Whether a progress bar should be provided.}
}
\description{
Find Address Candidates
}
