% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geocoder-obj.R
\name{geocode_server}
\alias{geocode_server}
\title{Create a GeocodeServer}
\usage{
geocode_server(url, token = arc_token())
}
\arguments{
\item{url}{the URL of a geocoding server.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[arcgisutils:auth_code]{auth_code()}}
or related function}
}
\value{
an object of class \code{GeocodeServer}.
}
\description{
Create an object of class \code{GeocodeServer} from a URL. This object
stores the service definition of the geocoding service as a list object.
}
\examples{
server_url <- "https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer"
geocode_server(server_url)
}
