% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-geocoders.R
\name{list_geocoders}
\alias{list_geocoders}
\alias{default_geocoder}
\title{List Available Geocoder Services}
\usage{
list_geocoders(token = arc_token())

default_geocoder(token = arc_token())
}
\arguments{
\item{token}{an object of class \code{httr2_token} as generated by \code{\link[arcgisutils:auth_code]{auth_code()}}
or related function}
}
\value{
a \code{data.frame}.
}
\description{
Evaluates the logged in user from an authorization token and returns
a data.frame containing the available geocoding services for the
associated token.
}
\details{
The \code{default_geocoder()}
}
\examples{
list_geocoders()
}
