% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse Geocode Locations}
\usage{
reverse_geocode(
  locations,
  crs = sf::st_crs(locations),
  ...,
  lang_code = NULL,
  feature_type = NULL,
  for_storage = FALSE,
  location_type = c("rooftop", "street"),
  preferred_label_values = c("postalCity", "localCity"),
  geocoder = default_geocoder(),
  token = arc_token(),
  .progress = TRUE
)
}
\arguments{
\item{locations}{an \code{sfc_POINT} object of the locations to be reverse geocoded.}

\item{crs}{the CRS of the returned geometries. Passed to \code{sf::st_crs()}.}

\item{...}{unused.}

\item{feature_type}{limits the possible match types returned. See Details.}

\item{for_storage}{default \code{FALSE}. Whether or not the results will be saved
for long term storage. See Details.}

\item{location_type}{default \code{"rooftop"}. Must be one of \code{"rooftop"} or \code{"street"}.}

\item{preferred_label_values}{default NULL. Must be one of \code{"postalCity"}
or \code{"localCity"}.}

\item{geocoder}{the url to the geocoder service.
Default provided by \code{default_geocoder()}.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[arcgisutils:auth_code]{auth_code()}}
or related function}

\item{.progress}{default \code{TRUE}. Whether a progress bar should be provided.}
}
\description{
Reverse Geocode Locations
}
\details{
\subsection{Feature Types}{

Valid \code{feature_type} values are \code{"StreetInt"}, \code{"DistanceMarker"},
\code{"StreetAddress"}, \code{"StreetName"}, \code{"POI"}, \code{"Subaddress"}, \code{"PointAddress"},
\code{"Postal"}, and \code{"Locality"}.

Intersection matches are only returned when \code{feature_types = "StreetInt"}.

See \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-reverse-geocode.htm#ESRI_SECTION3_1FE6B6D350714E45B2707845ADA22E1E}{documentation}.
}

\subsection{Storage}{

\strong{Very Important}

The argument \code{for_storage} is used to determine if the request allows you to
persist the results of the query. It is important to note that there are
contractual obligations to appropriately set this argument. \strong{You cannot save
or persist results} when \code{for_storage = FALSE} (the default).
}
}
